+++ crates/services/p2p/src/config.rs
@@ -9,27 +9,12 @@ use crate::{
     },
     peer_manager::ConnectionState,
 };
-use fuel_core_types::{
-    blockchain::consensus::Genesis,
-    secrecy::Zeroize,
-};
-use futures::{
-    future,
-    AsyncRead,
-    AsyncWrite,
-    Future,
-    FutureExt,
-    TryFutureExt,
-};
+use fuel_core_types::blockchain::consensus::Genesis;
+
 use libp2p::{
     core::{
         muxing::StreamMuxerBox,
         transport::Boxed,
-        upgrade::{
-            read_length_prefixed,
-            write_length_prefixed,
-        },
-        UpgradeInfo,
     },
     gossipsub::GossipsubConfig,
     identity::{
@@ -37,40 +22,42 @@ use libp2p::{
         Keypair,
     },
     mplex,
-    noise::{
-        self,
-        NoiseAuthenticated,
-        NoiseError,
-        NoiseOutput,
-        Protocol,
-    },
+    noise::{self,},
     tcp::{
         tokio::Transport as TokioTcpTransport,
         Config as TcpConfig,
     },
     yamux,
-    InboundUpgrade,
     Multiaddr,
-    OutboundUpgrade,
     PeerId,
     Transport,
 };
 use std::{
     collections::HashSet,
-    error::Error,
-    fmt,
-    io,
     net::{
         IpAddr,
         Ipv4Addr,
     },
-    pin::Pin,
     sync::{
         Arc,
         RwLock,
     },
     time::Duration,
 };
+
+use self::{
+    connection_tracker::ConnectionTracker,
+    fuel_authenticated::FuelAuthenticated,
+    fuel_upgrade::{
+        Checksum,
+        FuelUpgrade,
+    },
+    guarded_node::GuardedNode,
+};
+mod connection_tracker;
+mod fuel_authenticated;
+mod fuel_upgrade;
+mod guarded_node;
 
 const REQ_RES_TIMEOUT: Duration = Duration::from_secs(20);
 
@@ -80,16 +67,6 @@ const MAX_NUM_OF_FRAMES_BUFFERED: usize = 256;
 /// Adds a timeout to the setup and protocol upgrade process for all
 /// inbound and outbound connections established through the transport.
 const TRANSPORT_TIMEOUT: Duration = Duration::from_secs(20);
-
-/// Sha256 hash of ChainConfig
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Checksum([u8; 32]);
-
-impl From<[u8; 32]> for Checksum {
-    fn from(value: [u8; 32]) -> Self {
-        Self(value)
-    }
-}
 
 #[derive(Clone, Debug)]
 pub struct Config<State = Initialized> {
@@ -301,22 +278,26 @@ pub(crate) fn build_transport(
     let connection_state = ConnectionState::new();
 
     let transport = if p2p_config.reserved_nodes_only_mode {
+        let guarded_node = GuardedNode::new(&p2p_config.reserved_nodes);
+
+        let fuel_authenticated =
+            FuelAuthenticated::new(noise_authenticated, guarded_node);
+
         transport
-            .authenticate(GuardedNodeAuthentication::new(
-                noise_authenticated,
-                &p2p_config.reserved_nodes,
-            ))
+            .authenticate(fuel_authenticated)
             .apply(fuel_upgrade)
             .multiplex(multiplex_config)
             .timeout(TRANSPORT_TIMEOUT)
             .boxed()
     } else {
+        let connection_tracker =
+            ConnectionTracker::new(&p2p_config.reserved_nodes, connection_state.clone());
+
+        let fuel_authenticated =
+            FuelAuthenticated::new(noise_authenticated, connection_tracker);
+
         transport
-            .authenticate(FuelNodeAuthentication::new(
-                noise_authenticated,
-                &p2p_config.reserved_nodes,
-                connection_state.clone(),
-            ))
+            .authenticate(fuel_authenticated)
             .apply(fuel_upgrade)
             .multiplex(multiplex_config)
             .timeout(TRANSPORT_TIMEOUT)
@@ -326,320 +307,11 @@ pub(crate) fn build_transport(
     (transport, connection_state)
 }
 
-#[derive(Clone)]
-struct FuelNodeAuthentication<P, C: Zeroize, R> {
-    noise_authenticated: NoiseAuthenticated<P, C, R>,
-    reserved_nodes: HashSet<PeerId>,
-    connection_state: Arc<RwLock<ConnectionState>>,
-}
-
-impl<P, C: Zeroize, R> FuelNodeAuthentication<P, C, R> {
-    fn new(
-        noise_authenticated: NoiseAuthenticated<P, C, R>,
-        reserved_nodes: &[Multiaddr],
-        connection_state: Arc<RwLock<ConnectionState>>,
-    ) -> Self {
-        Self {
-            noise_authenticated,
-            reserved_nodes: reserved_nodes
-                .iter()
-                // Safety: as is the case with `bootstrap_nodes` it is assumed that `reserved_nodes` [`Multiadr`]
-                // come with PeerId included, in case they are not the `unwrap()` will only panic when the node is started.
-                .map(|address| PeerId::try_from_multiaddr(address).unwrap())
-                .collect(),
-            connection_state,
-        }
-    }
-}
-
-impl<P, C: Zeroize, R> UpgradeInfo for FuelNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo,
-{
-    type Info = <NoiseAuthenticated<P, C, R> as UpgradeInfo>::Info;
-    type InfoIter = <NoiseAuthenticated<P, C, R> as UpgradeInfo>::InfoIter;
-
-    fn protocol_info(&self) -> Self::InfoIter {
-        self.noise_authenticated.protocol_info()
-    }
-}
-
-impl<T, P, C, R> InboundUpgrade<T> for FuelNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo
-        + InboundUpgrade<T, Output = (PeerId, NoiseOutput<T>), Error = NoiseError>
-        + 'static,
-    <NoiseAuthenticated<P, C, R> as InboundUpgrade<T>>::Future: Send,
-    T: AsyncRead + AsyncWrite + Send + 'static,
-    C: Protocol<C> + AsRef<[u8]> + Zeroize + Send + 'static,
-{
-    type Output = (PeerId, NoiseOutput<T>);
-    type Error = NoiseError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_inbound(self, socket: T, info: Self::Info) -> Self::Future {
-        Box::pin(
-            self.noise_authenticated
-                .upgrade_inbound(socket, info)
-                .and_then(move |(remote_peer_id, io)| {
-                    let available_slot =
-                        self.connection_state.read().unwrap().available_slot();
-
-                    accept_any_node(
-                        &self.reserved_nodes,
-                        remote_peer_id,
-                        available_slot,
-                        io,
-                    )
-                }),
-        )
-    }
-}
-
-impl<T, P, C, R> OutboundUpgrade<T> for FuelNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo
-        + OutboundUpgrade<T, Output = (PeerId, NoiseOutput<T>), Error = NoiseError>
-        + 'static,
-    <NoiseAuthenticated<P, C, R> as OutboundUpgrade<T>>::Future: Send,
-    T: AsyncRead + AsyncWrite + Send + 'static,
-    C: Protocol<C> + AsRef<[u8]> + Zeroize + Send + 'static,
-{
-    type Output = (PeerId, NoiseOutput<T>);
-    type Error = NoiseError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_outbound(self, socket: T, info: Self::Info) -> Self::Future {
-        Box::pin(
-            self.noise_authenticated
-                .upgrade_outbound(socket, info)
-                .and_then(move |(remote_peer_id, io)| {
-                    let available_slot =
-                        self.connection_state.read().unwrap().available_slot();
-
-                    accept_any_node(
-                        &self.reserved_nodes,
-                        remote_peer_id,
-                        available_slot,
-                        io,
-                    )
-                }),
-        )
-    }
-}
-
-/// Wrapper over Noise protocol authentication.
-/// Used in case where the local node wants to limit
-/// who establishes the connection with it.
-/// During the Identity (PeerId) exchange, the node can check if the inbound connection
-/// comes from a whitelisted peer, in case it's not the connection is rejected.
-/// Same check is added to any outbound connections, just in case.
-#[derive(Clone)]
-struct GuardedNodeAuthentication<P, C: Zeroize, R> {
-    noise_authenticated: NoiseAuthenticated<P, C, R>,
-    reserved_nodes: HashSet<PeerId>,
-}
-
-impl<P, C: Zeroize, R> GuardedNodeAuthentication<P, C, R> {
-    fn new(
-        noise_authenticated: NoiseAuthenticated<P, C, R>,
-        reserved_nodes: &[Multiaddr],
-    ) -> Self {
-        Self {
-            noise_authenticated,
-            reserved_nodes: reserved_nodes
-                .iter()
-                // Safety: as is the case with `bootstrap_nodes` it is assumed that `reserved_nodes` [`Multiadr`]
-                // come with PeerId included, in case they are not the `unwrap()` will only panic when the node is started.
-                .map(|address| PeerId::try_from_multiaddr(address).unwrap())
-                .collect(),
-        }
-    }
-}
-
-/// Accepts a reserved node, or a non-reserved node if there is a free slot
-fn accept_any_node<T>(
-    reserved_nodes: &HashSet<PeerId>,
-    remote_peer_id: PeerId,
-    non_reserved_peers_allowed: bool,
-    io: NoiseOutput<T>,
-) -> future::Ready<Result<(PeerId, NoiseOutput<T>), NoiseError>> {
-    if reserved_nodes.contains(&remote_peer_id) || non_reserved_peers_allowed {
-        future::ok((remote_peer_id, io))
-    } else {
-        future::err(NoiseError::AuthenticationFailed)
-    }
-}
-
-/// Checks if PeerId of the remote node is contained within the reserved nodes.
-/// It rejects the connection otherwise.
-fn accept_reserved_node<T>(
-    reserved_nodes: &HashSet<PeerId>,
-    remote_peer_id: PeerId,
-    io: NoiseOutput<T>,
-) -> future::Ready<Result<(PeerId, NoiseOutput<T>), NoiseError>> {
-    if reserved_nodes.contains(&remote_peer_id) {
-        future::ok((remote_peer_id, io))
-    } else {
-        future::err(NoiseError::AuthenticationFailed)
-    }
-}
-
-impl<P, C: Zeroize, R> UpgradeInfo for GuardedNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo,
-{
-    type Info = <NoiseAuthenticated<P, C, R> as UpgradeInfo>::Info;
-    type InfoIter = <NoiseAuthenticated<P, C, R> as UpgradeInfo>::InfoIter;
-
-    fn protocol_info(&self) -> Self::InfoIter {
-        self.noise_authenticated.protocol_info()
-    }
-}
-
-impl<T, P, C, R> InboundUpgrade<T> for GuardedNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo
-        + InboundUpgrade<T, Output = (PeerId, NoiseOutput<T>), Error = NoiseError>
-        + 'static,
-    <NoiseAuthenticated<P, C, R> as InboundUpgrade<T>>::Future: Send,
-    T: AsyncRead + AsyncWrite + Send + 'static,
-    C: Protocol<C> + AsRef<[u8]> + Zeroize + Send + 'static,
-{
-    type Output = (PeerId, NoiseOutput<T>);
-    type Error = NoiseError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_inbound(self, socket: T, info: Self::Info) -> Self::Future {
-        Box::pin(
-            self.noise_authenticated
-                .upgrade_inbound(socket, info)
-                .and_then(move |(remote_peer_id, io)| {
-                    accept_reserved_node(&self.reserved_nodes, remote_peer_id, io)
-                }),
-        )
-    }
-}
-
-impl<T, P, C, R> OutboundUpgrade<T> for GuardedNodeAuthentication<P, C, R>
-where
-    NoiseAuthenticated<P, C, R>: UpgradeInfo
-        + OutboundUpgrade<T, Output = (PeerId, NoiseOutput<T>), Error = NoiseError>
-        + 'static,
-    <NoiseAuthenticated<P, C, R> as OutboundUpgrade<T>>::Future: Send,
-    T: AsyncRead + AsyncWrite + Send + 'static,
-    C: Protocol<C> + AsRef<[u8]> + Zeroize + Send + 'static,
-{
-    type Output = (PeerId, NoiseOutput<T>);
-    type Error = NoiseError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_outbound(self, socket: T, info: Self::Info) -> Self::Future {
-        Box::pin(
-            self.noise_authenticated
-                .upgrade_outbound(socket, info)
-                .and_then(move |(remote_peer_id, io)| {
-                    accept_reserved_node(&self.reserved_nodes, remote_peer_id, io)
-                }),
-        )
-    }
-}
-
-/// When two nodes want to establish a connection they need to
-/// exchange the Hash of their respective Chain Id and Chain Config.
-/// The connection is only accepted if their hashes match.
-/// This is used to aviod peers having same network name but different configurations connecting to each other.
-#[derive(Debug, Clone)]
-struct FuelUpgrade {
-    checksum: Checksum,
-}
-
-impl FuelUpgrade {
-    fn new(checksum: Checksum) -> Self {
-        Self { checksum }
-    }
-}
-
-#[derive(Debug)]
-enum FuelUpgradeError {
-    IncorrectChecksum,
-    Io(io::Error),
-}
-
-impl fmt::Display for FuelUpgradeError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            FuelUpgradeError::Io(e) => write!(f, "{}", e),            
-            FuelUpgradeError::IncorrectChecksum => f.write_str("Fuel node checksum does not match, either ChainId or ChainConfig are not the same, or both."),            
-        }
-    }
-}
-
-impl From<io::Error> for FuelUpgradeError {
-    fn from(e: io::Error) -> Self {
-        FuelUpgradeError::Io(e)
-    }
-}
-
-impl Error for FuelUpgradeError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match self {
-            FuelUpgradeError::Io(e) => Some(e),
-            FuelUpgradeError::IncorrectChecksum => None,
-        }
-    }
-}
-
-impl UpgradeInfo for FuelUpgrade {
-    type Info = &'static [u8];
-    type InfoIter = std::iter::Once<Self::Info>;
-
-    fn protocol_info(&self) -> Self::InfoIter {
-        std::iter::once(b"/fuel/upgrade/0")
-    }
-}
-
-impl<C> InboundUpgrade<C> for FuelUpgrade
-where
-    C: AsyncRead + AsyncWrite + Unpin + Send + 'static,
-{
-    type Output = C;
-    type Error = FuelUpgradeError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_inbound(self, mut socket: C, _: Self::Info) -> Self::Future {
-        async move {
-            // Inbound node receives the checksum and compares it to its own checksum.
-            // If they do not match the connection is rejected.
-            let res = read_length_prefixed(&mut socket, self.checksum.0.len()).await?;
-            if res != self.checksum.0 {
-                return Err(FuelUpgradeError::IncorrectChecksum)
-            }
-
-            Ok(socket)
-        }
-        .boxed()
-    }
-}
-
-impl<C> OutboundUpgrade<C> for FuelUpgrade
-where
-    C: AsyncRead + AsyncWrite + Unpin + Send + 'static,
-{
-    type Output = C;
-    type Error = FuelUpgradeError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
-
-    fn upgrade_outbound(self, mut socket: C, _: Self::Info) -> Self::Future {
-        async move {
-            // Outbound node sends their own checksum for comparison with the inbound node.
-            write_length_prefixed(&mut socket, &self.checksum.0).await?;
-
-            // Note: outbound node does not need to receive the checksum from the inbound node,
-            // since inbound node will reject the connection if the two don't match on its side.
-
-            Ok(socket)
-        }
-        .boxed()
-    }
+fn peer_ids_set_from(multiaddr: &[Multiaddr]) -> HashSet<PeerId> {
+    multiaddr
+        .iter()
+        // Safety: as is the case with `bootstrap_nodes` it is assumed that `reserved_nodes` [`Multiadr`]
+        // come with PeerId included, in case they are not the `unwrap()` will only panic when the node is started.
+        .map(|address| PeerId::try_from_multiaddr(address).unwrap())
+        .collect()
 }
